Add simplified osm/overpass importer
Based on feedback, going to ditch the old importer. Here's part of
its replacement.

This version doesn't require a config file and doesn't have many
switches. Takes 2 uris + either an area name or the -all-areas switch.
Options for -min-area, max-area also.

I tested this without actually writing anything... by using a really
y00ge -min-area and letting it skip everything. That seems to work
as expected. The rest of the code after that point should be the
same, so I expect that this works fine.