Add 2 new tools
* db-refresher

This is a tool that imitates the behavior of the
/api/config/reload?refresh=1 or all-spawnpoints=1 api calls.

This can be used to refresh the DB without reloading fletchling
and may be useful to ship separately in case of updates that
don't require restarting fletchling.

If satisfied with the results after running this tool, simply
reload fletchling with the /api/config call without any query
parameters.

* db-exporter

This is a standalone tool that exports nests in the DB. A
single nest can be exported as a geojson Feature or nests for
a particular area or all areas can be exported as a FeatureCollection.