Dynamic debug logging and spawnpoint age config (#49)
* Add endpoints to twiddle debug logging

This adds 2 new endpoints so that debug logging can
be turned on and off without restarting Fletchling.

* GET /debug/logging/on
* GET /debug/logging/off

This also fixes logging of the min_spawnpoints and area
config that was not cleaned up during a previous re-factor

* Add max_spawnpoint_age_days config option

This was hardcoded at 7 before and defaults to 7 now. Spawnpoints
older than this many days (not seen by Golbat) will not be
counted when computing spawnpoint counts for min_spawnpoints.